* Converted c-style casts to c++ casts (except dynamic_cast) where possible. The point here was to make it really easy to find the often non-portable reinterpret_casts to integer types and back, and make it easier to find and evaluate many other casts. * In many places casts were not necessary and were eliminated. * to_string and from_string moved out of SoarKernel's misc.h and in to shared project so that other projects could use it that didn't depend on the kernel. Removed kernel's misc.h and put the rest of its contents in utilities.cpp/h * Fixed an issue with Float and Int element (ClientSML) where const methods were actually const-casting away their own 'this' pointer and modifying themselves. * Changed some memory functions to use size_t when possible to make the code more portable * Cleaned up some of the rete's inline functions that were initially created by copy-and-pasting the old macros and therefore had many extra parens required in macros but not inline functions. The result is much easier to look at and understand. * All tests pass and things work fine on windows, testing unix/mac after this commit.

+++ SoarSuite/Core/SoarKernel/src/exploration.cpp
@@ -27,6 +27,7 @@
 
 #include "reinforcement_learning.h"
 #include "misc.h"
+#include "utilities.h"
 #include "instantiations.h"
 
 using namespace soar_TraceNames;
@@ -140,7 +141,7 @@ const long exploration_convert_parameter( agent *my_agent, const char *name )
 {
 	for ( int i=0; i<EXPLORATION_PARAMS; i++ )
 		if ( !strcmp( name, my_agent->exploration_params[ i ]->name ) )
-			return (const long) i;
+			return i;
 	
 	return EXPLORATION_PARAMS;
 }
@@ -716,7 +717,7 @@ preference *exploration_boltzmann_select( agent *my_agent, preference *candidate
 		 *  sum is negative, here that means a fractional probability
 		 */
 		double q_val = ( cand->numeric_value - q_diff );
-		total_probability += exp( (double) (  q_val / temp ) );
+		total_probability += exp( q_val / temp );
 		
 		/**
  		 * Let user know if adjusted q-value will overflow
@@ -736,7 +737,7 @@ preference *exploration_boltzmann_select( agent *my_agent, preference *candidate
 
 	for (cand = candidates; cand != NIL; cand = cand->next_candidate) 
 	{
-		current_sum += exp( (double) ( ( cand->numeric_value - q_diff ) / temp ) );
+		current_sum += exp( ( cand->numeric_value - q_diff ) / temp );
 		
 		if ( selected_probability <= current_sum )
 			return cand;
@@ -752,7 +753,7 @@ preference *exploration_epsilon_greedy_select( agent *my_agent, preference *cand
 {
 	preference *cand = 0;
 
-	double epsilon = exploration_get_parameter_value( my_agent, (const long) EXPLORATION_PARAM_EPSILON );
+	double epsilon = exploration_get_parameter_value( my_agent, static_cast< const long >( EXPLORATION_PARAM_EPSILON ) );
 
 	if ( my_agent->sysparams[ TRACE_INDIFFERENT_SYSPARAM ] )
 	{
